kernel/cpu/idt: Do not allow external 0x80 interrupts
svsm expects to receive a software interrupt as a result
of the INT 0x80 instruction. However, an external interrupt
on the same vector may also trigger the same codepath.

An external interrupt on vector 0x80 will currently be
interpreted as a 32-bit system call, and assuming that it
was a user context.

Panic on external interrupts on the vector.

To distinguish software interrupts from external ones,
svsm checks the APIC ISR bit relevant to the 0x80 vector.
For software interrupts, this bit will be 0.

Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>